requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pgvector
python-dotenv
google-generativeai
tiktoken
langchain-text-splitters
chromadb
